pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "guardian-one"
version = "0.1.0"
description = "Guardian One — multi-agent system for personal life management"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "cryptography>=42.0",
    "python-dotenv>=1.0",
    "schedule>=1.2",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
guardian = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
